// apps/client-v2/src/api/generated/models/Conductor.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type Conductor = {
    /**
     * @description The auto-generated id of the conductor
     * @type integer
    */
    id: number;
    /**
     * @description The name of the conductor
     * @type string
    */
    name: string;
    /**
     * @description The RUT of the conductor (CHILE)
     * @type string
    */
    rut: string;
};